+++ SwiftCheck/Test.swift
@@ -104,13 +104,19 @@ public func forAll<A : Arbitrary, B : Arbitrary, C : Arbitrary, D : Arbitrary, E
 
 /// Given an explicit generator and shrinker, converts a function to a universally quantified
 /// property.
-public func forAllShrink<A : Arbitrary>(gen : Gen<A>, shrinker: A -> [A], f : A -> Testable) -> Property {
+public func forAllShrink<A>(gen : Gen<A>, shrinker: A -> [A], f : A -> Testable) -> Property {
 	return Property(gen.bind { x in
 		return shrinking(shrinker, x, { xs  in
 			return counterexample("\(xs)")(p: f(xs))
 		}).unProperty
 	})
 }
+
+//public func forAll<A : Arbitrary>(pf : ([A] -> Testable)) -> Property {
+//	return forAllShrink(ArrayOf<A>.arbitrary().fmap({ $0.getArray }), { (x : [A]) -> [[A]] in
+//		return ArrayOf<A>.shrink(ArrayOf<A>(x)).map({ $0.getArray })
+//	}, pf)
+//}
 
 public func quickCheck(prop : Testable, name : String = "") {
 	quickCheckWithResult(stdArgs(name: name), prop)
